+++ src/POData/BatchProcessor/IBatchParser.php
@@ -1,0 +1,15 @@
+<?php
+namespace POData\BatchProcessor;
+
+use POData\BaseService;
+
+interface IBatchParser
+{
+    public function __construct(BaseService $host, $data);
+
+    public function process();
+
+    public function getResponse();
+
+    public function handleData();
+}