+++ src/POData/BatchProcessor/QueryParser.php
@@ -1,0 +1,11 @@
+<?php
+namespace POData\BatchProcessor;
+
+use \POData\OperationContext\ServiceHost;
+use Illuminate\Http\Request;
+use POData\BaseService;
+use POData\OperationContext\Web\Illuminate\IlluminateOperationContext;
+
+class QueryParser extends ChangeSetParser
+{
+}